requirements.txt
# Core inference
llama-cpp-python>=0.2.56
transformers>=4.40.0
accelerate>=0.30.0
bitsandbytes>=0.43.0

# ML utilities
torch>=2.1.0
numpy>=1.24.0
scipy>=1.11.0

# Knowledge graph
networkx>=3.1

# Data
datasets>=2.18.0

# Utilities
tqdm>=4.66.0
rich>=13.7.0
pydantic>=2.6.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0

# Paper / results
matplotlib>=3.8.0
pandas>=2.2.0
